# services/wallet-service/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "Core Wallet API",
        "termsOfService": "https://docs.wallet.example/terms",
        "contact": {
            "name": "Core Wallet Team",
            "email": "hoalh2@hdbank.com.vn"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/accounts/{acct_no}": {
            "get": {
                "description": "Account profile (no client PII).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Get account profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account number",
                        "name": "acct_no",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AccountResponse"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            },
            "patch": {
                "description": "Block / close / re-activate an account (status A | B | C).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Update account status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account number",
                        "name": "acct_no",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAccountStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AccountStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Business rule violation (e.g. illegal transition)",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
...
        "/v1/finance/transactions": {
            "get": {
                "description": "List ledger legs for an account, newest-first, keyset-paginated. Optional post_date range. Pass next_cursor as ?before_seq= for the next page.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Account statement (transaction list)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account number",
                        "name": "acct_no",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "post_date \u003e= (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "post_date \u003c= (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (1..200, default 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Keyset cursor: return rows with seq_no \u003c before_seq",
                        "name": "before_seq",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TxListResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Account not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/v1/finance/transactions/{tran_key}": {
            "get": {
                "description": "Return every ledger leg of one transaction, primary-leg first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "finance"
                ],
                "summary": "Transaction detail (all legs)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction internal key (tran_internal_key)",
                        "name": "tran_key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TxDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid transaction id",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/dto.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AccountResponse": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "acct_open_date": {
                    "type": "string"
                },
                "acct_role": {
                    "type": "string"
                },
                "acct_status": {
                    "type": "string"
                },
                "acct_type": {
                    "type": "string"
                },
                "actual_bal": {
                    "type": "string"
                },
                "available_bal": {
                    "description": "calc_bal = actual - restrained",
                    "type": "string"
                },
                "ccy": {
                    "type": "string"
                },
                "client_no": {
                    "type": "string"
                },
                "cr_blocked": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "last_tran_date": {
                    "type": "string"
                },
                "prev_day_bal": {
                    "type": "string"
                },
                "restrained_amt": {
                    "type": "string"
                },
                "restraint_present": {
                    "type": "string"
                },
                "shard_index": {
                    "type": "integer"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "dto.AccountStatusResponse": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "acct_status": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
...
        "dto.FieldError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "dto.ProblemDetails": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "RFC7807: human-readable detail (dynamic context)",
                    "type": "string",
                    "example": "available 50000 \u003c required 100000"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "errorCode": {
                    "description": "stable business error code (canonical contract)",
                    "type": "string",
                    "example": "INSUFFICIENT_FUNDS"
                },
                "errorMessage": {
                    "description": "stable human message for this code (i18n source, safe for end-user)",
                    "type": "string",
                    "example": "The account balance is not sufficient to cover this transaction"
                },
                "errors": {
                    "description": "field-level (Berlin Group / OBIE style)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FieldError"
                    }
                },
                "instance": {
                    "description": "RFC7807: request path",
                    "type": "string",
                    "example": "/v1/finance/transfer"
                },
                "internal_code": {
                    "description": "E#### for log/alert (§5)",
                    "type": "string",
                    "example": "E4022"
                },
                "iso20022_reason_code": {
                    "description": "ISO 20022 External Status Reason (§13.2)",
                    "type": "string",
                    "example": "AM04"
                },
                "retry": {
                    "$ref": "#/definitions/dto.RetryInfo"
                },
                "status": {
                    "description": "RFC7807: HTTP status code",
                    "type": "integer",
                    "example": 422
                },
                "timestamp": {
                    "description": "RFC 3339",
                    "type": "string",
                    "example": "2026-06-06T07:30:00Z"
                },
                "title": {
                    "description": "RFC7807: short human title",
                    "type": "string",
                    "example": "Insufficient funds"
                },
                "trace_id": {
                    "description": "= X-Request-Id",
                    "type": "string",
                    "example": "5f3b2c8e-1a2b-4c3d-9e8f-0a1b2c3d4e5f"
                },
                "transaction_status": {
                    "description": "pain.002 status (§13.3)",
                    "type": "string",
                    "example": "RJCT"
                },
                "type": {
                    "description": "RFC7807: URI to error doc",
                    "type": "string",
                    "example": "https://docs.wallet.example/errors/INSUFFICIENT_FUNDS"
                }
            }
        },
...
        "dto.RetryInfo": {
            "type": "object",
            "properties": {
                "after_ms": {
                    "type": "integer"
                },
                "retryable": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.TxDetailResponse": {
            "type": "object",
            "properties": {
                "leg_count": {
                    "type": "integer"
                },
                "legs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TxLegResponse"
                    }
                },
                "post_date": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "transaction_id": {
                    "description": "tran_internal_key",
                    "type": "integer"
                }
            }
        },
        "dto.TxEntryResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "balance_after": {
                    "type": "string"
                },
                "ccy": {
                    "type": "string"
                },
                "dr_cr": {
                    "type": "string"
                },
                "narrative": {
                    "type": "string"
                },
                "post_date": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "seq_no": {
                    "type": "integer"
                },
                "tran_type": {
                    "type": "string"
                },
                "transaction_id": {
                    "description": "= tran_internal_key",
                    "type": "integer"
                },
                "value_date": {
                    "type": "string"
                }
            }
        },
        "dto.TxLegResponse": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "amount": {
                    "type": "string"
                },
                "balance_after": {
                    "type": "string"
                },
                "ccy": {
                    "type": "string"
                },
                "dr_cr": {
                    "type": "string"
                },
                "narrative": {
                    "type": "string"
                },
                "post_date": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "seq_no": {
                    "type": "integer"
                },
                "tfr_seq_no": {
                    "type": "integer"
                },
                "tran_type": {
                    "type": "string"
                },
                "value_date": {
                    "type": "string"
                }
            }
        },
        "dto.TxListResponse": {
            "type": "object",
            "properties": {
                "acct_no": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "from": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TxEntryResponse"
                    }
                },
                "next_cursor": {
                    "description": "pass as ?before_seq= for the next page",
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateAccountStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "A",
                        "B",
                        "C"
                    ]
                }
            }
        },
...
    },
...
}